LogInStatusProviderを修正
・カスタムフックとそれを提供するコンポーネントの名称をより抽象度の高いものに変更
・利用側にとってはログインとログアウトを司る関数であることが伝わればいいため、switchLogInStatusをlogInとlogOutの二つに分割した
