+++ src/MemosSection.jsx
@@ -3,13 +3,13 @@ import Form from './Form.jsx'
 import MemoList from './MemoList.jsx'
 import { useMemos } from './useMemos.js'
 import './style/MemoList.css'
-import { useLogInStatus } from './LogInStatusProvider.jsx'
+import { useAuthentication } from './AuthenticationProvider.jsx'
 
 function MemosSection () {
   const [isVisibleForm, setIsVisibleForm] = useState(false)
   const [indexOfSelectedMemo, setIndexOfSelectedMemo] = useState(null)
   const memos = useMemos()
-  const { isLoggedIn, switchLogInStatus } = useLogInStatus()
+  const { isLoggedIn, logIn, logOut } = useAuthentication()
 
   const closeForm = () => {
     setIsVisibleForm(false)
@@ -25,7 +25,7 @@ function MemosSection () {
   return (
     <div className="listContainer">
       <div>
-        <button onClick={() => switchLogInStatus()}>
+        <button onClick={() => (isLoggedIn ? logOut() : logIn())}>
           {isLoggedIn ? 'ログアウト' : 'ログイン'}
         </button>
         <MemoList memos={memos} displayForm={displayForm} />
